Update README.md
This commit introduces the README.md file for 'Reborn to be Wilde', detailing the project's purpose, description, and initial guidance on installation and usage. It outlines our innovative approach to stylistic text transfer through model finetuning, aiming to emulate Oscar Wilde's writing style. The document also includes a preliminary section on contributing and recommends the MIT License for open usage and attribution.